# backend/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API Formation",
    "version": "1.0.0",
    "description": "API pour la plateforme de formation en ligne - Documentation Swagger avec tous les endpoints"
  },
  "servers": [
    {
      "url": "http://localhost:4004",
      "description": "Serveur local"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "Apprenant": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "utilisateurId": { "type": "integer" },
          "niveau": { "type": "string" }
        }
      },
...
      "Formation": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "titre": { "type": "string" },
          "description": { "type": "string" },
          "duree": { "type": "integer" },
          "niveau": { "type": "string" },
          "photo": { "type": "string" },
          "professeurId": { "type": "integer" },
          "categorie": { "type": "string" },
          "prix": { "type": "number" }
        }
      },
...
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/apprenants": {
      "get": {
        "summary": "Liste des apprenants",
        "tags": ["Apprenants"],
        "responses": {
          "200": {
            "description": "Liste des apprenants",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Apprenant"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Créer un apprenant",
        "tags": ["Apprenants"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "utilisateurId": { "type": "integer" },
                  "niveau": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Apprenant créé" }
        }
      }
    },
    "/apprenants/{id}": {
      "get": {
        "summary": "Récupérer un apprenant par ID",
        "tags": ["Apprenants"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "responses": {
          "200": { "description": "Apprenant trouvé" }
        }
      },
      "put": {
        "summary": "Mettre à jour un apprenant",
        "tags": ["Apprenants"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "niveau": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Apprenant mis à jour" }
        }
      },
      "delete": {
        "summary": "Supprimer un apprenant",
        "tags": ["Apprenants"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "responses": {
          "200": { "description": "Apprenant supprimé" }
        }
      }
    },
...
    "/formations": {
      "get": {
        "summary": "Liste des formations (Protégé - Filtré par rôle)",
        "tags": ["Formations"],
        "description": "Endpoint public - Accessible sans authentification",
        "responses": {
          "200": {
            "description": "Liste des formations",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Formation"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Créer une formation",
        "tags": ["Formations"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "titre": { "type": "string" },
                  "description": { "type": "string" },
                  "duree": { "type": "integer" },
                  "niveau": { "type": "string" },
                  "photo": { "type": "string" },
                  "professeurId": { "type": "integer" },
                  "categorie": { "type": "string" },
                  "prix": { "type": "number" }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Formation créée" }
        }
      }
    },
    "/formations/{id}": {
      "get": {
        "summary": "Récupérer une formation par ID (Public)",
        "tags": ["Formations"],
        "description": "Endpoint public - Accessible sans authentification",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "responses": {
          "200": { "description": "Formation trouvée" }
        }
      },
      "put": {
        "summary": "Mettre à jour une formation",
        "tags": ["Formations"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "titre": { "type": "string" },
                  "description": { "type": "string" },
                  "duree": { "type": "integer" },
                  "niveau": { "type": "string" },
                  "photo": { "type": "string" },
                  "categorie": { "type": "string" },
                  "prix": { "type": "number" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Formation mise à jour" }
        }
      },
      "delete": {
        "summary": "Supprimer une formation",
        "tags": ["Formations"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "responses": {
          "200": { "description": "Formation supprimée" }
        }
      }
    },
    "/formations/complete": {
      "post": {
        "summary": "Créer une formation complète avec chapitres et quiz",
        "tags": ["Formations"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "formation": {
                    "type": "object",
                    "properties": {
                      "titre": { "type": "string" },
                      "description": { "type": "string" },
                      "duree": { "type": "integer" },
                      "niveau": { "type": "string" },
                      "photo": { "type": "string" },
                      "professeurId": { "type": "integer" },
                      "categorie": { "type": "string" },
                      "prix": { "type": "number" }
                    }
                  },
                  "chapitres": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "titre": { "type": "string" },
                        "description": { "type": "string" },
                        "contenu": { "type": "string" },
                        "duree": { "type": "integer" },
                        "ordre": { "type": "integer" }
                      }
                    }
                  },
                  "quiz": {
                    "type": "object",
                    "properties": {
                      "titre": { "type": "string" },
                      "description": { "type": "string" },
                      "duree": { "type": "integer" },
                      "seuilReussite": { "type": "integer" }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Formation complète créée" }
        }
      }
    },
...
    "/progressions/apprenant/{apprenantFormationId}": {
      "get": {
        "summary": "Récupérer la progression d'un apprenant",
        "tags": ["Progressions"],
        "parameters": [
          {
            "name": "apprenantFormationId",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "responses": {
          "200": { "description": "Progression récupérée" }
        }
      }
    },
...
    "/formations/demandes/en-attente": {
      "get": {
        "summary": "Récupérer les formations en attente de validation (Admin)",
        "tags": ["Formations - Admin"],
        "description": "Accessible uniquement aux administrateurs",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "description": "Formations en attente de validation" }
        }
      }
    },
    "/formations/{id}/valider": {
      "put": {
        "summary": "Valider une formation (Admin)",
        "tags": ["Formations - Admin"],
        "description": "Accessible uniquement aux administrateurs - Valide une formation en attente",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "responses": {
          "200": { "description": "Formation validée" }
        }
      }
    },
    "/formations/{id}/rejeter": {
      "put": {
        "summary": "Rejeter une formation (Admin)",
        "tags": ["Formations - Admin"],
        "description": "Accessible uniquement aux administrateurs - Rejette une formation en attente",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "responses": {
          "200": { "description": "Formation rejetée" }
        }
      }
    }
  }
}